refactor(db): inline connection factory into database manager
- Replace DatabaseConnectionFactory with internal Hikari config
- Initialize database tables on datasource creation
- Simplify configuration using AskimoHome for paths
- Disable Docker tests in CLI release workflow for consistency
