+++ shared/src/main/kotlin/io/askimo/core/db/DatabaseManager.kt
@@ -4,12 +4,14 @@
  */
 package io.askimo.core.db
 
+import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
 import io.askimo.core.chat.repository.ChatDirectiveRepository
 import io.askimo.core.chat.repository.ChatFolderRepository
 import io.askimo.core.chat.repository.ChatMessageRepository
 import io.askimo.core.chat.repository.ChatSessionRepository
 import io.askimo.core.chat.repository.ConversationSummaryRepository
+import io.askimo.core.util.AskimoHome
 import java.sql.Connection
 import javax.sql.DataSource
 
@@ -25,12 +27,52 @@ class DatabaseManager private constructor(
     useInMemory: Boolean = false,
 ) : AutoCloseable {
 
-    private val hikariDataSource: HikariDataSource =
-        DatabaseConnectionFactory.createSQLiteDataSource(
-            databaseFileName = databaseFileName,
-            initializeDatabase = ::initializeTables,
-            useInMemory = useInMemory,
-        )
+    private val hikariDataSource: HikariDataSource = createSQLiteDataSource(
+        databaseFileName = databaseFileName,
+        useInMemory = useInMemory,
+    )
+
+    /**
+     * Creates a HikariDataSource for a SQLite database file in the Askimo home directory.
+     *
+     * @param databaseFileName The name of the database file (e.g., "askimo.db")
+     * @param useInMemory If true, creates an in-memory database (useful for testing)
+     * @return A configured HikariDataSource
+     */
+    private fun createSQLiteDataSource(
+        databaseFileName: String,
+        useInMemory: Boolean,
+    ): HikariDataSource {
+        val jdbcUrl = if (useInMemory) {
+            // Use in-memory database for tests to avoid file locking issues on Windows
+            // Use file URI with cache=shared so all connections in the pool share the same database
+            // Each repository instance will have its own isolated in-memory database
+            "jdbc:sqlite:file:memdb_${System.nanoTime()}?mode=memory&cache=shared"
+        } else {
+            val dbPath = AskimoHome.base().resolve(databaseFileName).toString()
+            "jdbc:sqlite:$dbPath"
+        }
+
+        val config = HikariConfig().apply {
+            this.jdbcUrl = jdbcUrl
+            driverClassName = "org.sqlite.JDBC"
+            maximumPoolSize = if (useInMemory) 1 else 10 // Single connection for in-memory
+            minimumIdle = if (useInMemory) 1 else 2
+            connectionTimeout = 30000
+            idleTimeout = 600000
+            maxLifetime = 1800000
+            connectionInitSql = "PRAGMA foreign_keys = ON;"
+            addDataSourceProperty("cachePrepStmts", "true")
+            addDataSourceProperty("prepStmtCacheSize", "250")
+            addDataSourceProperty("prepStmtCacheSqlLimit", "2048")
+        }
+
+        return HikariDataSource(config).also { ds ->
+            ds.connection.use { conn ->
+                initializeTables(conn)
+            }
+        }
+    }
 
     /**
      * The datasource for obtaining database connections.
